refactor: move value_converter to gin
