+++ filenames.gni
@@ -481,7 +481,8 @@ filenames = {
     "shell/common/gin_converters/net_converter.cc",
     "shell/common/gin_converters/net_converter.h",
     "shell/common/gin_converters/std_converter.h",
-    "shell/common/gin_converters/value_converter_gin_adapter.h",
+    "shell/common/gin_converters/value_converter.cc",
+    "shell/common/gin_converters/value_converter.h",
     "shell/common/gin_helper/arguments.cc",
     "shell/common/gin_helper/arguments.h",
     "shell/common/gin_helper/callback.cc",
@@ -516,10 +517,6 @@ filenames = {
     "shell/common/mouse_util.h",
     "shell/common/mac/main_application_bundle.h",
     "shell/common/mac/main_application_bundle.mm",
-    "shell/common/native_mate_converters/v8_value_converter.cc",
-    "shell/common/native_mate_converters/v8_value_converter.h",
-    "shell/common/native_mate_converters/value_converter.cc",
-    "shell/common/native_mate_converters/value_converter.h",
     "shell/common/node_bindings.cc",
     "shell/common/node_bindings.h",
     "shell/common/node_bindings_linux.cc",
@@ -541,6 +538,8 @@ filenames = {
     "shell/common/promise_util.cc",
     "shell/common/skia_util.h",
     "shell/common/skia_util.cc",
+    "shell/common/v8_value_converter.cc",
+    "shell/common/v8_value_converter.h",
     "shell/renderer/api/context_bridge/render_frame_context_bridge_store.cc",
     "shell/renderer/api/context_bridge/render_frame_context_bridge_store.h",
     "shell/renderer/api/atom_api_context_bridge.cc",
