+++ shell/browser/api/atom_api_system_preferences_mac.mm
@@ -27,7 +27,7 @@
 #include "shell/browser/ui/cocoa/NSColor+Hex.h"
 #include "shell/common/deprecate_util.h"
 #include "shell/common/gin_converters/gurl_converter.h"
-#include "shell/common/gin_converters/value_converter_gin_adapter.h"
+#include "shell/common/gin_converters/value_converter.h"
 #include "ui/native_theme/native_theme.h"
 
 namespace gin {
