+++ shell/browser/api/gpuinfo_manager.cc
@@ -11,6 +11,7 @@
 #include "content/public/browser/browser_thread.h"
 #include "gpu/config/gpu_info_collector.h"
 #include "shell/browser/api/gpu_info_enumerator.h"
+#include "shell/common/gin_converters/value_converter.h"
 
 namespace electron {
 
@@ -46,7 +47,7 @@ void GPUInfoManager::ProcessCompleteInfo() {
   // We have received the complete information, resolve all promises that
   // were waiting for this info.
   for (auto& promise : complete_info_promise_set_) {
-    promise.Resolve(*result);
+    promise.ResolveWithGin(*result);
   }
   complete_info_promise_set_.clear();
 }
@@ -86,7 +87,7 @@ void GPUInfoManager::FetchBasicInfo(
     util::Promise<base::DictionaryValue> promise) {
   gpu::GPUInfo gpu_info;
   CollectBasicGraphicsInfo(&gpu_info);
-  promise.Resolve(*EnumerateGPUInfo(gpu_info));
+  promise.ResolveWithGin(*EnumerateGPUInfo(gpu_info));
 }
 
 std::unique_ptr<base::DictionaryValue> GPUInfoManager::EnumerateGPUInfo(
