+++ shell/common/gin_converters/value_converter.h
@@ -2,12 +2,10 @@
 // Use of this source code is governed by the MIT license that can be
 // found in the LICENSE file.
 
-#ifndef SHELL_COMMON_NATIVE_MATE_CONVERTERS_VALUE_CONVERTER_H_
-#define SHELL_COMMON_NATIVE_MATE_CONVERTERS_VALUE_CONVERTER_H_
+#ifndef SHELL_COMMON_GIN_CONVERTERS_VALUE_CONVERTER_H_
+#define SHELL_COMMON_GIN_CONVERTERS_VALUE_CONVERTER_H_
 
-#include "native_mate/converter.h"
-
-#include "base/optional.h"
+#include "gin/converter.h"
 
 namespace base {
 class DictionaryValue;
@@ -15,7 +13,7 @@ class ListValue;
 class Value;
 }  // namespace base
 
-namespace mate {
+namespace gin {
 
 template <>
 struct Converter<base::DictionaryValue> {
@@ -35,29 +33,6 @@ struct Converter<base::Value> {
                                    const base::Value& val);
 };
 
-template <typename T>
-struct Converter<base::Optional<T>> {
-  static bool FromV8(v8::Isolate* isolate,
-                     v8::Local<v8::Value> val,
-                     base::Optional<T>* out) {
-    if (val->IsNull() || val->IsUndefined()) {
-      return true;
-    }
-    T converted;
-    if (Converter<T>::FromV8(isolate, val, &converted)) {
-      return true;
-    }
-    out->emplace(converted);
-    return true;
-  }
-  static v8::Local<v8::Value> ToV8(v8::Isolate* isolate,
-                                   const base::Optional<T>& val) {
-    if (val.has_value())
-      return Converter<T>::ToV8(val.value());
-    return v8::Undefined(isolate);
-  }
-};
-
 template <>
 struct Converter<base::ListValue> {
   static bool FromV8(v8::Isolate* isolate,
@@ -67,6 +42,6 @@ struct Converter<base::ListValue> {
                                    const base::ListValue& val);
 };
 
-}  // namespace mate
+}  // namespace gin
 
-#endif  // SHELL_COMMON_NATIVE_MATE_CONVERTERS_VALUE_CONVERTER_H_
+#endif  // SHELL_COMMON_GIN_CONVERTERS_VALUE_CONVERTER_H_
