+++ shell/common/v8_value_converter.cc
@@ -2,7 +2,7 @@
 // Use of this source code is governed by the MIT license that can be
 // found in the LICENSE file.
 
-#include "shell/common/native_mate_converters/v8_value_converter.h"
+#include "shell/common/v8_value_converter.h"
 
 #include <map>
 #include <memory>
