+++ shell/common/v8_value_converter.h
@@ -2,8 +2,8 @@
 // Use of this source code is governed by the MIT license that can be
 // found in the LICENSE file.
 
-#ifndef SHELL_COMMON_NATIVE_MATE_CONVERTERS_V8_VALUE_CONVERTER_H_
-#define SHELL_COMMON_NATIVE_MATE_CONVERTERS_V8_VALUE_CONVERTER_H_
+#ifndef SHELL_COMMON_V8_VALUE_CONVERTER_H_
+#define SHELL_COMMON_V8_VALUE_CONVERTER_H_
 
 #include <memory>
 
@@ -74,4 +74,4 @@ class V8ValueConverter {
 
 }  // namespace electron
 
-#endif  // SHELL_COMMON_NATIVE_MATE_CONVERTERS_V8_VALUE_CONVERTER_H_
+#endif  // SHELL_COMMON_V8_VALUE_CONVERTER_H_
