+++ shell/renderer/electron_api_service_impl.cc
@@ -14,7 +14,7 @@
 #include "mojo/public/cpp/system/platform_handle.h"
 #include "shell/common/atom_constants.h"
 #include "shell/common/gin_converters/blink_converter.h"
-#include "shell/common/gin_converters/value_converter_gin_adapter.h"
+#include "shell/common/gin_converters/value_converter.h"
 #include "shell/common/heap_snapshot.h"
 #include "shell/common/node_includes.h"
 #include "shell/common/options_switches.h"
